move manufacturer models into separate file
